Add proxy as a valid request option
Sometimes you need to use an outbound proxy to connect to other services. To do this you would have to set the proxy header.

This is useful when hosting an app on a Paas or similar where you don't have static IPs, and would need to connect/proxy to services that use IP whitelisting.