Update README.md with information about timeout
We had a problem with sporadic failures with the default settings due to a slow backend server. It turns out the default timeout value (5 seconds) had to be increased, however, the default value was nowhere to be found in the documentation. This PR adds documentation on the timeout property and its default value so that other users don't have to spend time looking of this value in the source code.