Handle original query for URLs without ? character
When using `originalQuery: true` option, if the URL does not have a `?` character, the regex `/^.+\?/` causes the entire request URL to be set as the `search` parameter. This change checks for the presence of a `?` character in the string, and sets an empty string as the search parameter if not present.